Use metaprogramming to define GoodTimes functions
Since GoodTimes requires Meta, we can't use functions in GoodTimes inside Meta.
This is why Seconds and Now have been moved out, and the @types put in Now.

`GoodTimes.now` isn't set up by macro, so to make it available straight from
GoodTimes, there is a delegation to Now. Unfortunately documentation doesn't
follow along in a delegation, and so it's duplicated.

Seconds could have been delegated, but then we run into the documentation issue
again. This time, we can instead define those methods with the macro. It's a
bit weird, since some of them they essentially duplicate the methods in
`GoodTimes.Seconds`.

I've so far been unable to set the parameter names to the macro defined
functions dynamically. They show up in documentation. Where it would read
`seconds_ago(seconds)`, it now reads `seconds_ago(integer)`.

The macro takes a singular version of the time unit. The plural is derived in
the most naive way: sticking an "s" on the end.