Remove duplication in tests
Almost all tests repeat the same pattern. Specify expected time difference,
compute actual difference, assert equality. I've extracted this to a function,
passing in the datetimes (or datetime, if we're comparing to `now`), and the
expected difference.

I find that it doesn't hurt readability much, even on slightly long lines, and
it makes the tests easier to scan.